rewrites as LayeredGraph during propagation, if length.assigned()
This is actually a little tricky. Adds OViewTraits<View> class, modelled on the exitsing VarTraits<Var> class.
The template just has a public typedef for OViewTraits::Var, which is a variable type based on the view type used to instantiate the template class.
Use this type to create the VarArgArray which is use in the call to lgp_post() during rewriting.