restores rewriting
Without rewriting, or with rewriting happening after the backwards pass, approximately 10% of the solutions to the test case were missed (using N then A branching).
If rewriting happens when the length is fixed, but BEFORE the backwards pass in prop(), then all solutions are found.
This would seem to imply that backwards is broken for the fixed length case (at least sometimes?).