adds compression in final layer, major overhaul of extend()
IMPORTANT: This version is, at a first glance, much worse than the previous one.

Have a dfa-map form the (currently) fiunal layer back to the dfa states.
Use that map when consulting the distance table, which is still in terms of dfa states.
Major overhaul of extend(), much more careful about when to enter layers:
	layer[0] initialized to a single state
	forward pass layers n->length.min()-1
	backward pass length.min() -> 0,
		but careful not to resubscribe variables for layers under n
	compression layers n->length.min()
Also change how n_edges is updated on extend: now just count the edges in every layer (since we are traversing them all in the backward pass anyway).