BUG FIX: missing bracket around else block in "finality" check
This resulted in around 50% of solutions getting lost.
When i_deg == 0 or distance+n > length.max, should set o_deg = 0.
However, added a check to see if that was a change, without making it a block.
Result was that the check executed correctly, but the instruction executed always!

Honsetly, it's a miracle that any solutions made it past that.